feat: add SwaggerConfig, SecurityUtils. add new features to AdminService

+++ src/main/java/com/matheus/task_manager_api/service/AdminService.java
@@ -1,5 +1,7 @@
 package com.matheus.task_manager_api.service;
 
+import com.matheus.task_manager_api.enums.Role;
+import com.matheus.task_manager_api.exception.ForbiddenActionException;
 import com.matheus.task_manager_api.mapper.TaskMapper;
 import com.matheus.task_manager_api.mapper.UserMapper;
 import com.matheus.task_manager_api.dto.TaskResponse;
@@ -8,6 +10,7 @@ import com.matheus.task_manager_api.entity.User;
 import com.matheus.task_manager_api.exception.UserNotFoundException;
 import com.matheus.task_manager_api.repository.TaskRepository;
 import com.matheus.task_manager_api.repository.UserRepository;
+import com.matheus.task_manager_api.security.SecurityUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -36,8 +39,18 @@ public class AdminService {
     }
 
     public void deleteUser(Long id) {
+        User admin = SecurityUtils.getAuthenticatedUser();
+
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException(id));
+
+        if (user.getId().equals(admin.getId())) {
+            throw new ForbiddenActionException("You cannot delete your own account");
+        }
+
+        if (user.getRole().equals(Role.ADMIN)) {
+            throw new ForbiddenActionException("You cannot delete another admin");
+        }
 
         userRepository.delete(user);
     }
